requirements.txt
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0
##deepspeed>=0.12.0
wandb>=0.16.0
Pillow>=9.0.0
requests>=2.25.0
tqdm>=4.65.0
numpy>=1.21.0
bitsandbytes>=0.41.0